changes made for to pay amount fifth commit

+++ app/src/main/java/com/example/wallet/Api/PaidTransaction.java
@@ -11,7 +11,7 @@ public class PaidTransaction {
     public Object transferDate;
     public String paymentId;
     public String amount;
-    public double transactionFee;
+    public float transactionFee;
     public double settlementFee;
     public double amountPayable;
     public double processignPaidByIns;
